Extract Item.Export
